Skeleton for implementing a bucket
