+++ gstorage/checks.py
@@ -2,33 +2,10 @@
 """
 Check that the module is registered with required settings
 """
-
-from os import environ
-
-from django.conf import settings
 from django.core import checks
 
 from .constants import REQUIRED_SETTINGS
-
-
-def get_config(key):
-    """
-    Get the value of config from settings. Else get it from environment.
-
-    Args:
-        key: A string indicating the key to config, for ex - GOOGLE_APPLICATION_CREDENTIALS
-
-    Returns:
-        The value of the key if present, else None
-    """
-    try:
-        return getattr(settings, key)
-    except AttributeError:
-        try:
-            return environ[key]
-        except KeyError:
-            pass
-    return None
+from .utils import get_config
 
 
 def check_gstorage_params(**kwargs):
